Exporting a Static site creates empty js and css folders
Modified the slashes in Directory paths to utilize the default file
delimiter for the system.
This seemed to fix the issue.
During my debugging process I also identified on two different windows
instances, one running coldfusin 9 with IIS and the other running Railo
4 with Apache, when creating a new site, the js and css folders copied
files over with a zero length.  This will be the next bug for me to
tackle.
My first commit.
Thanks for the great CMS Blue River!